requirements.txt
requests
beautifulsoup4
asyncpg
aiogram==3.*
python-dotenv
spacy